Split content of `CnbClient` to multiple files
