+++ jobs/Backend/Task/src/ExchangeRateUpdater/Cnb/CnbError.cs
@@ -1,0 +1,5 @@
+﻿namespace ExchangeRateUpdater.Cnb;
+
+// 💡 I would normally put this declaration next to interface (defining signature of functionality),
+//    but we could possibly introduce child classes to distinguish between different errors 
+public class CnbError;
